requirements.txt
altgraph==0.17.5
packaging==26.0
pefile==2024.8.26
pygame-ce==2.5.6
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
python-i18n==0.3.9
pywin32-ctypes==0.2.3
